templated DataBackend test
